fixed rendering

+++ src/main/java/com/project/worker3d/Point3d.java
@@ -1,9 +1,9 @@
 package com.project.worker3d;
 
 public class Point3d {
-    public float x, y, z = 0;
+    public double x, y, z = 0;
 
-    public Point3d(float x, float y, float z) {
+    public Point3d(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
